这一次在我mac 机器上bench ，上一个 pr 是在我的主机上测试的，所以 参考基准不一样。
master branch: 

```bash
is_chinese("扁担宽，板凳长，扁担想绑在板凳上。")                                                                                             
                        time:   [83.674 ns 84.042 ns 84.506 ns]
Found 9 outliers among 100 measurements (9.00%)
  4 (4.00%) high mild
  5 (5.00%) high severe

                                                                                                                                              is_chinese("ss扁担宽，板凳长，扁担想绑在板凳上。")                                         
                        time:   [2.0029 ns 2.0085 ns 2.0145 ns]
Found 5 outliers among 100 measurements (5.00%)
  2 (2.00%) high mild
  3 (3.00%) high severe

                                                                                                                                              is_chinese("扁担宽，板凳长，扁担想绑在板凳上。ss")                                         
                        time:   [43.542 ns 43.863 ns 44.326 ns]
Found 7 outliers among 100 measurements (7.00%)
  3 (3.00%) high mild
  4 (4.00%) high severe

isChinese(chars1000) true")                                                                             
                        time:   [4.6331 us 4.6485 us 4.6646 us]
Found 5 outliers among 100 measurements (5.00%)
  1 (1.00%) high mild
  4 (4.00%) high severe

is_chinese("isChinese(chars1001) false")                                                                             
                        time:   [2.5277 us 2.5390 us 2.5559 us]
Found 9 outliers among 100 measurements (9.00%)
  5 (5.00%) high mild
  4 (4.00%) high severe
```
perf branch: 
```bash
is_chinese("扁担宽，板凳长，扁担想绑在板凳上。")                                                                                              
                        time:   [38.533 ns 38.819 ns 39.152 ns]
                        change: [-54.813% -54.023% -53.260%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 5 outliers among 100 measurements (5.00%)
  2 (2.00%) high mild
  3 (3.00%) high severe

                                                                                                                                              is_chinese("ss扁担宽，板凳长，扁担想绑在板凳上。")                                         
                        time:   [4.8136 ns 4.8368 ns 4.8620 ns]
                        change: [+138.40% +140.61% +142.54%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 5 outliers among 100 measurements (5.00%)
  2 (2.00%) high mild
  3 (3.00%) high severe

                                                                                                                                              is_chinese("扁担宽，板凳长，扁担想绑在板凳上。ss")                                         
                        time:   [41.423 ns 41.788 ns 42.235 ns]
                        change: [-5.8544% -4.5882% -3.3184%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 14 outliers among 100 measurements (14.00%)
  3 (3.00%) high mild
  11 (11.00%) high severe

isChinese(chars1000) true")                                                                             
                        time:   [2.1482 us 2.1632 us 2.1816 us]
                        change: [-54.895% -54.039% -53.332%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 13 outliers among 100 measurements (13.00%)
  6 (6.00%) high mild
  7 (7.00%) high severe

is_chinese("isChinese(chars1001) false")                                                                             
                        time:   [2.1280 us 2.1559 us 2.1968 us]
                        change: [-17.167% -15.932% -14.731%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 9 outliers among 100 measurements (9.00%)
  3 (3.00%) high mild
  6 (6.00%) high severe
```
可以看到，除了 ascii 在字符串开头的情况，其均有大幅度提升，全是中文的情况下 提升大约两倍