@IWANABETHATGUY 2021-10-19
去掉 第一层循环是因为 rust 的 chars 迭代器 overhead 相比 bytes 会高很多，这里单独做一次循环不是很划算，去掉将在大部分情况提速很多。